biquadtest: Add some diagnostic logic for a weird problem with 32-bit compiler
This is code description of some debug in progress.

Running on F14/i686 there is an odd problem with the FPU build that y1
is not consistent after the call to tt_biquad_step. As can be seen
the memcmp() proves that b1 and b2 are binary identical but the result
of executing tt_biquad_step() yeilds different results.

The problem only occurs at -O1 or faster.

I don't yet understand why.

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>