libtt/jackwriter: First attempt at jack integration
At the moment we retain strict interface compatibility with the other
audio writers. Specifically the audio system blocks the thread that
tries to issue the samples until it is ready. jack wasn't really designed
to work like this so the code "feels" a bit fragile (although I've not
yet found any problems).

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>