libtt/drummachine: Fix signed division bug
TT_DRUMMACHINE_VOICE_LENGTH is used in division expressions on signed
numbers. It must therefore be signed to ensure these expressions don't
perform unsigned division (thus not propagating the sign bit).

Curiously this bug cannot be observed on 64-bit systems where the
expression converts to 64-bit unsigned division and a later downcast to
32-bit restores the sign bit!

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>