librfn/fibre: Keep track of scheduler tainting
This is mostly to reduce the need to error paths in application code.
We can permit applications to ignore the result of fibre_run_atomic()
(for example) providing all they need for error handling is a kernel
taint to be added.

Eventually this mechanism will call a (weakly linked) taint handler to
allow us to perform a reset or deploy some other aggressive recovery
mechanism.

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>