librfn/fibre: Fix silly list ordering bug in the scheduler.
A bit embarrassing this one... thankfully the asserts in the list
handling picked it up very easily when it triggered on real silicon.

We also modify the behavior of fibre_self() both to make the test suite
more powerful and also to allow the main loop to implement task runtime
accounting.

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>