fibre: posix: Impose a 50ms upper limit on sleep times.
On a microcontroller the fibre scheduler is intended to go idle using a
wait-for-interrupt mechanism and will leave idle both at the end of the
sleep period or if there is any asynchronous activity on the device.

POSIX doesn't provide this type of mechanism but we can emulate it by
waking every 50ms and checking that there's nothing for the scheduler to
do. Note that this is just responsive for "human speed" interfacing but
impractical for most other purposes.

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>