console: Add explicit character get/put support
console_getch() can be used by console commands to interact with the
user. console_eval() allows characters to be programmatically injected
into the command processor.

The later is a bit of a hack since it doesn't wait for the command
processor to be idle (i.e. injection only when no pending characters
have been inserted by the user) but is sufficient to implement automatic
boot time scripting.

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>