messageq: Fix messageq_empty()
Currently messageq_empty() has the side effect of *consuming* the
message we are peeking at. This is a pretty severe bug so we fix it
and add a test case to cover this API.

We also take the opportunity to inline the implementation of
messageq_empty() as part of some optimization work within the scheduler.
We also fix the include order at the top of the C file (so ensure we
automatically test that the messageq.h header file is self-sufficient.

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>