benchmark: Reduce test cycles from 100 to 50
As we add more benchmarks the benchmark is taking longer to run. The
cheapest way to speed up test execution is to reduce the number of
iterations of the suite we run and push though the min/mean/max
calculations.

We also modify the test harness to report total execution time in a
humane way.

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>